--- src/api/mod.rs
@@ -1,6 +1,0 @@
-pub mod activity;
-pub mod admin;
-pub mod build_config;
-pub mod builder;
-pub mod validator;
-pub mod verification;
